Add showcase README with real interface captures and finish project docs
Document the problem, quick start, architecture, metrics, tests, and
limitations, with screenshots and an animated flow recording captured
from the running app. Align the upload limit with the 5 MB CSV contract,
store per-inquiry counts as integers, rewrite the demo guide for the new
interface, and keep local working notes out of version control.

+++ sql/cohort.sql
@@ -16,7 +16,7 @@ SELECT i.*,
     coalesce(b.bookings, 0) AS bookings,
     coalesce(b.cancellations, 0) AS cancellations,
     coalesce(b.no_shows, 0) AS no_shows,
-    coalesce(b.completed_jobs, 0) AS completed_jobs,
+    coalesce(b.completed_jobs, 0)::BIGINT AS completed_jobs,
     coalesce(b.revenue_cents, 0)::BIGINT AS revenue_cents,
     coalesce(b.revenue_cents, 0) / 100.0 AS revenue_usd
 FROM inquiries i LEFT JOIN booking_totals b USING (inquiry_id)
